+++ osu_bot/API/Parameters/BeatmapInfoQueryParameters.cs
@@ -5,7 +5,7 @@ namespace osu_bot.API.Parameters
 {
     public class BeatmapInfoQueryParameters : IQueryParameters
     {
-        public int BeatmapId { get; set; }
+        public long BeatmapId { get; set; }
 
         public string GetQueryString() => $"https://osu.ppy.sh/api/v2/beatmaps/{BeatmapId}";
     }
